pyproject.toml
[project]
name = "rez"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "fastmcp>=2.12.2",
    "httpx>=0.28.1",
    "mcp[cli]>=1.13.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.30.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "certifi>=2025.8.3",
    "beautifulsoup4>=4.13.5",
]
